+++ src/com/kaylerrenslow/armaDialogCreator/arma/header/HeaderArrayItem.java
@@ -7,4 +7,8 @@ import org.jetbrains.annotations.NotNull;
  @since 03/19/2017 */
 public interface HeaderArrayItem {
 	@NotNull HeaderValue getValue();
+
+	default boolean arrayItemEquals(@NotNull HeaderArrayItem o) {
+		return o == this || this.getValue().equals(o.getValue());
+	}
 }