+++ src/com/kaylerrenslow/armaDialogCreator/arma/header/HeaderAssignment.java
@@ -9,4 +9,9 @@ public interface HeaderAssignment {
 	@NotNull String getVariableName();
 
 	@NotNull HeaderValue getValue();
+
+	default boolean equalsAssignment(@NotNull HeaderAssignment o) {
+		return o == this || getVariableName().equals(o.getVariableName()) && getValue().equals(o.getValue());
+	}
+
 }