+++ src/com/kaylerrenslow/armaDialogCreator/arma/header/HeaderValue.java
@@ -7,4 +7,8 @@ import org.jetbrains.annotations.NotNull;
  @since 03/19/2017 */
 public interface HeaderValue {
 	@NotNull String getContent();
+
+	default boolean equalsValue(@NotNull HeaderValue o) {
+		return this.getContent().equals(o.getContent());
+	}
 }