+++ tests/com/kaylerrenslow/armaDialogCreator/arma/header/largeTestRoot_configs.h
@@ -1,0 +1,15 @@
+class Configuration
+{
+	class Reputation
+	{
+		max = 500; //The max amount of Reputation any non admin player can have
+
+		class Earnings //how much Reputation can be earned in each situation
+		{
+			kill_inf = 1;   //killed a single enemy infantry
+		};
+
+		class Penalties;
+
+	};
+};